refactor: Info class 사용 추가
